rename file for ThisLaunchFileDir substitution
